moved estimate to ftqc directory
