+++ resource_estimation/ftqc/estimate_test.py
@@ -15,7 +15,7 @@ from math import pi
 import cirq
 import pytest
 import resource_estimation.ftqc.architecture as arch
-import resource_estimation.estimate as est
+import resource_estimation.ftqc.estimate as est
 import resource_estimation.ftqc.lattice_surgery_primitives as lsp
 from numpy import isclose
